refactor 24.10.07 domain 분리
