+++ src/test/java/racingcar/RacingCarTest.java
@@ -9,6 +9,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Race;
 
 class RacingCarTest {
     private Race race;
